Removes "graceful degradation is fault-tolerant"
Graceful degradation is *broken* by design. Progressive enhancement starts from a position of "your site works on every single user agent". Graceful degradation starts from a position of "your site works on a handful of the tens of thousands of user agents that exist" … and there is no possibility of you ever catering for more than a handful of them.