switches case of consts
Standard practice is to just use regular camelCase. If you go with the const as default approach then you can assume that everything is immutable and you don’t need a naming convention (not that a naming convention would necessarily be a good idea anyway, unless you have some way to enforce it. They're too easy for people to ignore/forget/be-unaware-of, and then you have the potential for confusion)