Be more welcoming (#275)
Given this is now the top Google result for "frontend playbook", I think less prominent mention of ownership is appropriate. It now could be seen as the discipline's de-facto standard frontend playbook.

Advertising/treating it as such, may be beneficial to us (as it will present a more welcoming face), and to others (as they may feel more comfortable utilising it if they can refer to it as a standard, rather than simply a company's preferred way of doing things - which it is as well, but … you know, it's also utilised by many others).

Lots of justification for deleting two words, but that's how I work innit.