Tweaking the rationale for browser support (#286)
Was unclear what "this" referred to in the introductory sentence, so have provided something that reads well without context.

Nice mention of the Pareto principle btw! Only a slight tweak to this part, as the support costs being low is more of an inevitable result of the approach, than a motivator for it. Support costs should be _appropriate_.